file-upload-service-main/requirements.txt
Django
djangorestframework
pandas
openpyxl
